controller
add controller has function for cart include 
checks if the productid and increment variables have been set. The condition 
is used to determine the action to be perform. In this instance, we increment 
searches through the collection of items using the product id and returns an array of rowIds that match the search criteria. The rowId is a unique identify of an item in the Cart and it is auto generated. It is used to update items in the Cart collection